requirements.txt
roboflow
ultralytics
tensorflow
opencv-python-headless
tqdm
Pillow